Add new util method for dumping node hashes
